Lazily import globus_sdk components for speed
On python3.7+, when using globus-sdk/main (where lazy imports are
currently available), this reduces the time to run `globus -h` by
approximately 50%. Similar improvements apply to completions.

In order to drive and support the refactoring process, two tooling
changes were made first:

- a new flake8 local plugin which forbids `from globus_sdk import ...`
- a test which insists that `from globus_cli import main` does not
  import `requests`

For the most part, we switch to `import globus_sdk`-style imports and
that takes care of everything. Some type annotations need to be
quoted.

There are a few notable exceptions, some of which could possibly be
improved in the future:

- `globus_cli.services` contains specialized client subclasses.
  Because class definition requires import evaluation, we defer any
  internal imports of these components using the standard deferral
  idiom of "import within a function"
- Registered error handlers used exception classes from the SDK at
  import time (decorator args). This now accepts strings which get
  pulled from `globus_sdk` to defer the import to a later step.
- `globus_sdk.GroupRequiredSignupFields` was used to define a Choice
  argument. This forces an import, so we need to switch to a string
  list (can add tests to ensure its exhaustive and accurate)
- `globus_sdk.scopes` is okay to import. This is quite by accident,
  but as long as an SDK component imports without dependency chains
  and doesn't use `requests`, we can use it. This is nice for the
  LoginManager