Add "minimal" hidden 'timer create' command (#672)
* Initial draft of 'globus timer create transfer'

Introduce the `globus timer create` command as a hidden command.
In support of this new command, several broader changes are included:

- some `globus transfer` options are now declared as shared option
  decorators for reuse on the timer create command
- a new `TimedeltaType(ParamType)` is created
- the login manager is adjusted to get the requisite timer scopes
- `globus transfer --batch` handling is now abstracted into a helper
  for reuse on the timer command

* Allow sync_level_option to be applied directly

Allow `@sync_level_option` as the decorator, rather than
`@sync_level_option()`

Also, move `sync_level_option` to a dedicated module for easier
maintenance. In the process of doing this rewrite, adjust
`globus_cli.parsing` to use local relative imports for brevity.

* Add tests for TimedeltaType and allow whitespace

Upon review, it's easy enough to add support for whitespace in the
expression for parsing timedeltas. This will allow quoted usages with
spaces. Add tests which exercise the TimedeltaType in all of its
various usages.

* Add tests for timer create transfer

- some basic end-to-end tests ensure the command can run
- explicit tests for various bad usages which are caught
- make the timezone resolution a function so that it can be tested in
  unit tests more thoroughly, and patched over more easily for
  relevant functional tests

* Adjust timedelta: reject bare integers

Bare int support is ambiguous, it is not costly for users to specify
the unit, and it exposes us to odd inputs like `1d4`, which is either
the damage roll on a normal dagger, someon really wanting "a day plus
four seconds", or an accidental truncation.

* Remove unused constant

* Refactor sharing of Transfer task submit options

Move these all into a dedicated module, and make sure more are shared
between `globus transfer` and `globus timer create transfer`.

In cases where the timer option name appeared to be more correct in
some way (e.g. `--encrypt-data` over `--encrypt`, which matches the
API fieldname), both names are preserved for the `globus transfer`
command only, by means of aliases.