Fix synopsis generation in reference generator
When a synopsis is provided explicitly, there should be no codegen
which combines it with the command name. This was still being done,
which appeared in the form of a misplaced backtick in some docs, but
made the fix broken as well (missing backticks in other docs).

The correct solution is to do codegen if `adoc_synopsis` is not set,
not if it is set, and avoid this combination of the two.