Update to mypy==0.982 and add update script (#683)
1. Copy the SDK version bump script and use it to write a mypy update
   script which tweaks tox.ini and .pre-commit-config.yaml
2. Apply the script update (latest mypy version is now in use)
3. Adjust one annotation to pass under the new mypy

Regarding (3), the failure stems from an annotation being more
specific than is really needed. An extra type variable is being used
and bound as a parameter to another type variable -- probably this
should be supported by mypy, but perhaps there's a reason that it
should not. In any case, the typevar is not really needed since the
entire value in question is being passed through a `cast` call after
the usage site. Therefore, the wayward type variable can be safely
replaced with Any.

Regarding (1), this may be an unnecessary script being added. However,
it's a nice and low-key way of exploring the potential space for an
auto-update process which ensures the mypy version stays consistent.
Next step: should we hook this script in to generate automatic PRs
which bump `mypy` in both locations?