Improve defaults for 'collection create mapped'
`storage_gateway_id` and `base_path` are now specified via options
rather than positionals. These have two advantages:
- order independent (too many positional arguments becomes confusing)
- we can specify defaults

Because we make a strong habit of avoiding required options, it is
desirable to have some default behaviors.
`--base-path=/` is simple, but `--storage-gateway` (note that it does
not end in `-id`) is more complex.

After discussion, we're going to have it default to "the only gateway
on the endpoint", and leave any more sophisticated behaviors as future
work. Future possibilities include

    # selecting by type
    --storage-gateway=posix

    # selecting by name
    --storage-gateway=mygateway

    # selecting by any of the above with prefix mapping
    --storage-gateway="type:posix"
    --storage-gateway="name:mygateway"