Invert the default for disallow_untyped_defs and enumerate non-compliant modules explicitly (#722)
* Invert application of disallow_untyped_defs

Rather than default-False and having to set this to True where
applicable, default-True and set to False where applicable.
Elimination of `*` rules is left as future work.

* mypy.ini: enumerate modules in exception_handling

Convert '.*' to individual module names. Rather than also listing the
top-level name (__init__.py), it is equally simple to add the relevant
annotations.

* mypy.ini: enumerate modules in login_manager

* mypy.ini: enumerate modules in parsing.param_types

* mypy.ini: enumerate modules in parsing.shared_options

* mypy.ini: enumerate modules in services.transfer

* mypy.ini: enumerate modules in 'commands'

Unlike various other manual changes, this was autogenerated by
removing the rule for `commands.*`, running mypy, and putting the
output through a pipeline to generate lines of config.

This ensures that disallow_untyped_defs=false is only set for existing
modules, and those modules are named explicitly in all cases.