Add click annotation checker and begin using it (#738)
* Add click annotation checker and begin using it

The annotation checker has specialized knowledge of how click params
map back to types, and customizations to handle the param types defined
in the CLI.

It encodes logic for mapping the parameter list for command objects
back to expected sets of annotated arguments. It then checks
annotations *in a runtime, test context* against those expectations
and raises errors if there are detected deviations.

The new helper is applied to several recently annotated commands and
has the following basic results:
- `--notify` needs to be moved to a custom param type so that it can be
  understood by the annotation checker
- `--batch` was annotated incorrectly in `globus transfer`; missing
  `| None` in the annotation

Importantly, `--notify` heralds a potential sweeping change, that any
custom callback may require special handling. As an alternative, we
could introduce custom subclasses of Option and Argument which allow
us to specify known types as part of the runtime specification.

The tests only run on py3.10+ in order to allow use of
`types.UnionType` and have other useful features.

* Add tests for `--notify` and fix a bug

The latest change which converted the `--notify` callback to a type
introduced a regression. Missing `--notify` would be passed as `None`,
rather than the expected `{}`. To fix, introduce a callback which
simply converts `None` to `{}` explicitly. The logic for how
"non-nullable options" in the annotation checker needs updating to
handle this corectly, and the new tests are using the callback as a
simple way to exercise its behavior.

* Use future import to fix test collection on py<3.9

* Refine click param type checker and add self-tests

This changeset refines the click parameter type checker to be more
readable and simpler. Primarily, it factors out the logic of
converting a click.Parameter object into a type. Additionally, it
reduces the walks of parameters down to a single loop for improved
readability.

As part of the refactor, unit tests for the param checker itself are
introduced. These help guarantee that the checker runs correctly on
various "simple" cases. In the course of this testing, some minor bugs
with the checker were revealed:
- `multiple` params should always be `tuple[X, ...]`, they do not
  default to `None`
- `multiple` params can take the form of an argument with `nargs=-1`,
  account for this

This applies a fix which was noted in review, to union-ize the
multiple types seen under a multiple option and include them in the
tuple type.

An additional fix to the `globus transfer` command is here identified:
the type for `exclude` does not include `None`. Setting a default of
`None` has no effect. The default is `()`. The annotation is
corrected, and a relevant line of source is also fixed.

* Add a custom completer to `--notify` option

`globus {transfer,delete} --notify` now gets a custom completer which
provides for various reasonable usages with the following behavior:

  - `--notify <TAB>` -> `["on", "off", "succeeded", "failed", "inactive"]`
  - `--notify o<TAB>` -> `["on", "off"]`
  - `--notify failed,<TAB>` -> `["failed,succeeded", "failed,inactive"]`
  - `--notify failed,s<TAB>` -> `["failed,succeeded"]`

Also make the following changes:
- remove `if value is None` check, a remnant of click v7.x resilient
  parsing mode
- correct the docstring to document the output type correctly

Include unit tests to exercise the new completion behavior.

* Refactor NotificationParamType for clarity

The main body of `NotificationParamType`, the `convert()` method, was
hard to read because it created too many variables in a single scope
and rebound the type of one of those variables.

This rewrite (proposed by @kurtmckee) makes the main body clearer by
avoiding changing the type of `value`, and dropping the `on` and `off`
bools. It also improves the behavior when an invalid option is given
by calling it out by name in the error message.

Additionally, fix type annotations on the notify param type and on the
none-to-bool callback.

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* Simplify `--notify` shell complete; fewer paths

There are fewer conditional and branching points in this rewrite of
the functionaltiy, inspired heavily by comments from @kurtmckee .

A minor simplification: collect results from `split(",")` using
left-hand-side `*args` capturing.

More significantly: condense all of the logic after the `"," in str`
check down to almost a single path. The logic is now trying to build a
set of possible completions for the last post-comma fragment,
*including* the case when that fragment is `""`. There are a couple of
corner cases which need careful handling, but overall the logic is
simpler and less nested.

* Improve test coverage of notify param type

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>