Fix test response registration time (#755)
The registration of our response data can take a non-negligible amount
of time (YAML parsing is slow and there are many files).

Unfortunately, structuring this as an autouse session fixture puts
that within the timing for the execution of the first test. In concert
with pytest-timeout being applied, this can in rare cases push a test
over the timeout. (e.g. If run on a machine under heavy load.)

The issue can be dodged entirely by moving the registration of our
response data into pytest_configure or pytest_sessionstart, which
happen before any tests are running and being timed.