Strictly enforce static scope requirements (#761)
The LoginManager now checks to see if the scope requirements for
statically declared scopes are met, according to the data in token
storage. When scope requirements are not met, the login is not
considered valid for the `requires_login` decorator and related
methods. As a result, when new scope requirements are added to the
CLI, users may be prompted to re-login in order to gain the newly
desired scope.

This ensures that newly added scope requirements are enforced, which
means that the applicaiton has fewer potential states to reason about
when scopes are added.

The STATIC_SCOPES map has added new sections but has not updated the
scope list for any section since the release of globus-cli v3.0.0.
Therefore, we can be confident that no users on modern versions of the
CLI will be impacted by this change on its own.