Testsuite cleanup: remove LoginManager._TEST_MODE (#760)
There was a testing-aware path in the source which allowed a callout
to be enabled/disabled with a class variable. Although this might be
nice from the perspective of the testsuite, it isn't really
necessary. It's very easy to replace this with a more clearly named
monkeypatch usage in the testsuite, and use `MonkeyPatch.undo()` to
reverse the patch in the one test which needs to do this.