Fix handling of `globus api -Q` for multi-params (#759)
A parameter passed multiple times with different values should be
encoded and sent with all of the values, not just the last one. This
constitutes as switch from tracking these as `dict[str, str]` to
`dict[str, list[str]]`. Tests confirm that this correctly encodes
parameters like a Transfer ls `filter`.

Some minor refactoring to the tests for the `globus api` command makes
this simpler to write.