Annotate task commands and fix bugs in `globus task list` filters (#767)
* Apply missing annotations to 'task' commands

Remove the mypy disallow_untyped_defs=false config for these commands,
get them passing, and run them through the click annotation checker
test to ensure that the annotations are accurate.

This highlighted that `globus task list` had code which looked for
multiple opts to be specifically `list` values when they are actually
`tuple`s. The relevant helper also operates over lists, so the
behavior is corrected by checking if the value is a Sequence type,
and the annotations now match this. No tests for the runtime change
are included here, but they should be added as future work.

* Add tests for `globus task list`

This also revealed a bug, fixed herein, in the behavior of the
request_time and completion_time filters, which were being assembled
in the wrong order.

* Fix missing future anotations imports