Add dynamic dependencies to the Timer scope and enforce them (#765)
* Add dynamic dependencies to Timer scope

Timer requires that we build a dynamic dependency for the
TransferAP->Transfer dependency chain. Therefore, we need sufficient
tooling to build this dependency chain *and* to build the dependency
chain with data_access scopes at the leaves.

To support this, a `scopes` module is added to the `login_manager`
subpackage and it contains a helper for building Timer scopes as well
as a couple of useful constants.

This also introduces the notion of "contract versions" to the login
manager, but without useful enforcement logic at present. The idea of
the contract version is that it tracks when our expectations/needs for
a scope change in a way which we don't believe can be detected without
an explicit flag.

* Add scope "contract version" enforcement

Even with our recent refinements to the login manager to capture scope
requirements, there are still scope configuration changes which cannot
be captured by our detection logic.

In order for the CLI to actively attempt to detect these conditions,
we have two options:
- make more aggressive use of the Consents API
- add some internal tracking so that we know when our scopes change

The existing checks we've added for scope strings are still good and
useful for the majority of cases. However, for dynamic scope
dependencies, we need one of these mechanisms. Relevantly, the Timer
scope needs to be used with dynamic dependencies attached to it, and
this cannot be detected by the mechanisms we have today.

Using the Consents API would mean that when checking if the user is
logged in, we need to make a call out to Auth. This extra roundtrip is
part of a growing speed penalty on "launch" of a command that we would
ideally avoid. Using local storage is marginally more complex but much
faster because we just need to make a roundtrip to disk to get info on
what we last requested.

Rather than storing the dynamic dependencies or other specific data
which might not handle other scenarios in the future, this introduces
a generic "scope contract version" number, which is declared starting
at 1 (with existing tokens implicitly under contract version 0). We
store this in the config table of tokenstorage and mark all tokens
with our current version number in a `resource_server: integer`
mapping. The login checks can then simply compare the known contract
version against data from disk and thereby get a comparison between
our current scope requirements and the requirements at the time that
the user did a login flow for that token.

In order to implement this, the tokenstorage wrapping module needs to
define a custom SQLiteAdapter subclass. And to maintain our deferred
import performance tuning in the CLI, the subclass needs to be defined
dynamically inside of a helper. Once that is taken care of, however,
the implementation is "relatively straightforward".

The CURRENT_SCOPE_CONTRACT_VERSION is a constant initialized with a
value of 1, and is what will be written to disk.

* Rewrite the form of token/scope requirements

Define a new class, `_CLIScopeRequirements` (internal) and a singleton
instance of it, `CLI_SCOPE_REQUIREMENTS` which extract the notions
previously encoded by `LoginManager.STATIC_SCOPES` and several other
attributes of the LoginManager class.

Because this changes how we address resource servers by known "nice"
names, it touches almost every command in the CLI with an update to
their login requirements. Most of these changes are purely cosmetic.

The LoginManager now depends upon the CLI_SCOPE_REQUIREMENTS and
interacts with it, so the problems of one and the other are somewhat
separated.

CLI_SCOPE_REQUIREMENTS can address information by "nice names" like
"transfer" and by full resource server names. Each service is listed
with their min_contract_version for the scope contract versions idea.
The data within the mapping is structured using a TypedDict, which
gives us some level of type-safety in our accesses, and the patching
for login manager tests can simply rewrite the full suite of data
using ordinary dicts to get desired test-time semantics out of the
system.

* Make the set of service short-names a Literal

This is shared across the whole CLI and allows us to validate more of
our usages.

* Update src/globus_cli/login_manager/manager.py

Co-authored-by: derek-globus <113056046+derek-globus@users.noreply.github.com>

* Add a test to enforce ServiceNameLiteral matching

This test ensures that ServiceNameLiteral matches the keys in
CLI_SCOPE_REQUIREMENTS exactly.

* Test current contract version and make it Final

Also, refactor the test fixtures a little to avoid repeatedly
calling `MonkeyPatch.undo` to undo an autouse fixture.

* Require `requires_login` args to be sorted

Add a new check to our internal flake8 plugin to enforce this, and
conform.

---------

Co-authored-by: derek-globus <113056046+derek-globus@users.noreply.github.com>