Add publishing workflows + version "smudger" (#775)
* Add version smudging script for dev releases

This is a script which adds a `devX` version number to the end of the
current version string. It takes a `--revert` flag which will return
the version number to a 'X.Y.Z' version assuming that the smudger was
used to modify the version in the first place.

This tool is meant to support automated testpypi releases using
modified version numbers.

* Create new CI workflows for automatic publishing

These are copied from globus-sdk and kept nearly identical.
However, the workflow for publishing to test-pypi runs the version
smudging script as an intermediate step so that we can rerun the
workflow as needed if there's ever a problem with a release.

* Refine dev version "smudge script" per review

Avoid using 'epoch' which is a special term in python version
specifiers. Simply call it `--number` and state it as a dev release
number. Set the type to `int`, and compute `dev{N}` before calling the
same helper as before.

Rename the whole script to `set_dev_version.py` to more clearly state
its purpose/behavior.