Add server timing header ASCII graph with hidden flag control (#784)
* First draft of opt-in Server-Timing parsing

This pulls in a parser and hooks it into `display(r)` when `r` is an
SDK response object. Printing targets stderr and uses colors to make
things pretty.

* Control server-timing display with hidden option

`--show-server-timing` is a hidden option which will enable the
rendering of Server-Timing headers. This replaces the env var which
was used in the first draft.

To support this, a new bool field gets attached to the state object
and a new hidden option goes into the common opts to flip that bool
flag. A helper in the "context" module for 'globus_cli.termio' can
inspect the state object for this flag.

Also, fixup some type annotations and comments which could be
trivially improved.

* Refactor ServerTiming parse and add initial tests

A framework is now in place to support multiple versions of the
Server-Timing header spec, although the only one which is supported is
still the 2017 version in which there is no `dur` key for the
duration, and the format is primarily `name=duration`.

Future parsers could be added in the future and wrapped such that they
are equally well supported. At that point, it would be possible to try
multiple parses to support multiple different spec versions if
desired.

* Apply suggestions from code review

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add a whitespace stripping test for Server-Timing

And add a call to `strip()` in the source to make the test pass.

This *may* allow for technically non-compliant header values to
parse, but this is arguably a desirable outcome.

---------

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>