Rewrite JSONStringOrFile to accept filenames without the 'file:' prefix and have a simpler implementation (#790)
* Add JSONStringOrFileV2 to improve JSON parsing

This is a new param type which is intended to replace JSONStringOrFile
at all usage sites. As a first step in the process, introduce the new
type, authored to be a fresh implementation with very similar
behavior, and copy the existing JSONStringOrFile tests to produce a
compatibility test which verifies that the new object behaves
similarly.

As a key difference, the new implementation does not require the
`file:` prefix, but will treat it as a disambiguator which means "do
not try a JSON parse of this data".

The order of precedence for the parse is:
- `file:` prefix (if present)
- file loading
- JSON parse (if no such file)

Errors are treated somewhat carefully, to try to ensure that the JSON
parse will not hide a missing file error and misreport it. However,
this requires ultimately attempting to detect whether or not the input
was intended to be JSON or a filename. This is done with an
intentionally simplistic heuristic: check if the first char is `{`.

As an improvement to the interface, `JSONStringOrFileV2` does not
return the raw loaded JSON data, but places it into a dataclass
wrapper which pairs the loaded data with the filename. Although many
usages will not benefit from this explicit treatment, it makes it
possible to give back an error on malformed data which refers to the
file containing said data.

Additionally, JSONStringOrFileV2 handles `-` to indicate stdin,
opening up additional possible usages.

* Convert collection update to JSONStringOrFileV2

* Update user-credential from-json parsing

Update these to use JSONStringOrFileV2. Additionally, now that the
argumetns are being annotated, add these commands to the click
annotation checker.

Now that the data type is known on these data to be any valid JSON,
mypy is able to detect and flag non-dict data as a possibility. To
resolve this, add new runtime checks which confirm that the data is
object-shaped.
New tests exercise the runtime checks.

* Convert 'flows start' to JSONStringOrFileV2

Update the parsing to use JSONStringOrFileV2, update the docstring to
show a filename without the `file:` prefix, and update the command
internals slightly to handle the new parsed data object.

This catches a potential type error in which non-dict data could have
been parsed and would have been passed through. Now that the types are
more precise, it is flagged by mypy as a type checker error. To
resolve, a new check is added to confirm that the input data is
object-shaped.
A new test exercises this check.

* Remove JSONStringOrFile

Remove the older implementation of JSONStringOrFile. The only place
where it still appeared was in the param types tests, where it can be
replaced now with the JSONStringOrFileV2 type.

* Rename "JSONStringOrFileV2" to "JSONStringOrFile"

This completes the replacement process, renaming the new type to match
the name of the old type.

* Use JSONStringOrFile in search commands

This replaces the use of `click.File('r')` combined with explicit
calls to `json.load`. The end result is that we are using the
`JSONStringOrFile` helper in all locations where it is valid to apply,
and the datatypes within the relevant search commands
(delete-by-query, query, and ingest) are knowable for mypy.

As a result of these improvements, we need some additional changes.
Runtime checks for non-dict data (as in other uses of ParsedJSONData),
and in one case a widening type cast for `list[str]` to `list[JSON]`.

* Add changelog for JSONStringOrFile rewrite

* Fix coverage gaps from JSONStringOrFile rewrite

The rewrite and error handling it introduced created several minor
code coverage gaps. Close these up with proper tests.

* Update src/globus_cli/parsing/param_types/json_strorfile.py

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* Improve description of JSON object data

Per review, improve the phrasing to make things more concise and
clear.

---------

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>