Replace use of the SDK paginator table with 'wrap' (#826)
The paginator table attached to client objects provides a convenient
syntax for accessing paginators which wrap the relevant underlying
client methods.

However, it is not presently possible to explain it to type checkers
that this is a type-manipulating proxy over the client class. The
paginator table is just one valid syntax for accessing a Paginator
object built from the method being paginated. The paginator building
is actually done by `Paginator.wrap`, which inspects the method for
special pagination attributes and dispatches to its relevant
subclass.

The major advantage of `Paginator.wrap()` is that it preserves type
annotation information correctly, returning a paginator object
parametrized over the return type of the underlying method.

As a result, we should prefer `Paginator.wrap` for mypy-checked code,
and the current SDK docs even have a note to this effect. [1]

In order to ensure the move is done everywhere and holds into
the future, add a pygrep hook which looks for access to a `.paginated`
attribute. If this runs into false positives in the future, it could
be replaced with an AST-based lint or something else more
sophisticated, but on the current codebase it is fully accurate.

The change can be verified to be effective by introducing unuspported
arguments and seeing that `mypy` catches them. For example:

    ```diff
    -    res = paginator(task_id).items()
    +    res = paginator(task_id, foo=bar).items()
    ```

is flagged:

    ```
    $ mypy src/
    src/globus_cli/commands/task/show.py:71: error: Unexpected keyword argument "foo"  [call-arg]
    ```

Such a check was not possible via the paginator-table interface (the
`.paginated` attribute).

[1]: https://globus-sdk-python.readthedocs.io/en/stable/core/paging.html#typed-paginators-with-paginator-wrap