Add type annotations to the login manager subpackage (#829)
* Apply annotations to most 'login_manager' modules

* Type annotate LoginManager and accommodate typing

Most of the annotation changes have no broad impact.
However, one change in particular is nefariously tricky:
`requires_login`.

This is a decorator factory which needs to be annotated in a way to
indicate that it adds an argument. ParamSpec and typing do not support
annotating the behavior we have today, in which the decorator adds the
`login_manager` keyword argument. The closest possible phrasing is to
use ParamSpec+Concatenate to indicate that we pass it as the first
positional argument. Therefore, *change the runtime logic to use this
pattern* and apply the ParamSpec-based annotations needed to validate
this in all commands.

In all cases where `login_manager` is passed via `requires_login`,
move it to be the only positional argument and annotate all other
command parameters as keyword-only, using the `*` separator. This also
uniformizes several commands which had drifted away from the preferred
style of keyword-only argument declarations.

The final detail on this change is that `mypy` has recently
experienced a regression in the handling of ParamSpecs under complex
scenarios like decorator factories. There is at least one potential
fix in progress, but it is unclear when it might release as it does
not appear to be ready for the upcoming v1.5 release.
However, `mypy` *does* correctly understand the actual type of such a
decorator factory -- this can be tested with `reveal_type` -- so we
can use this feature but with a type-ignore in a relatively
critical-looking location (the return value of the decorator
factory).

In addition to these changes, applying this change to
`session show` resulted in some other annotation issues being flagged
by mypy, so those are fixed here as well.