Convert Transfer Timer consent check to use trees (#833)
Rather than doing explicit matching down the desirable path in the
`timer create transfer` command, produce a MutableScope tree with the
correct path and match it against the consent graph (forest) returned
by the consents API.

Because the tree-like structure of MutableScope objects mirrors those
which will come from the experimental scope parser in the SDK, this
conversion serves as preliminary verification that we can match a
subforest against the consent graph and find matching or non-matching
paths.

For the initial implementation, the match is purely boolean in nature,
finding only whether or not a scope was present. In the future, this
could be made to return a list of mismatches or other meaningful diff
data.

In service of these changes, the CustomAuthClient class now supports
calling the consents API and returning a ConsentForestResponse object.
The ConsentForestResponse is where the tree matching functionality
resides, although it could easily be moved.