Convert flows list --orderby to colon-split choice
`Tuple(Choice, Choice)` is not gracefully handled by click's
autocompleter, meaning that one of our goals in terms of
user-experience would not be met if we pursued this path.

Replace that usage with an intelligent colon-separated tuple type,
which can provide literal types for annotation checking but looks like
a `click.Choice` for the purpose of completions.