Fix handling of gcs_manager_url
Rather than referring to the first element of the server list by
hostname, fetch the `"gcs_manager_url"` field from the get_endpoint
payload.

This update matches a recommendation from the Transfer API for
deducing this value and fixes a bug in which an empty server list
would make commands fail on this lookup.

This has been tested on several modern collections and endpoints (at
time of testing, the test collections are running GCS v5.4.66), and
appears to work correctly under both the before-and-after CLI
versions, as both fields are populated.

Given that the test fixture data lists v5.4.10, presumably these
fixtures are far out of date, and we can trust that `gcs_manager_url`
will be populated in all cases.

In the worst case, we could revert this change back to using the
server list as a fallback.

The URL needs to be formatted with the `/api` suffix whereas this was
not needed before. This accounts for a behavior of the GCSClient class
in which an address specified as a hostname is formatted as
`https://{address}/api`, whereas one which is formatted as a URL is
not formatted. So the SDK was previously appending `/api` and
prepending `https://`, but now it will not be making any change to the
address string.