Convert param type checker to click-type-test
This is a lift-and-shift of our existing functionality in this repo
into a public and open-source alpha project.

The logic being isolated means that globus-cli no longer bears
responsibility for testing it internally either, so the tests which
were a self-test of the param type checker have been migrated into the
new project and eliminated.