Introduce `--orderby` for `globus ls`
Using the colon-delimited choice paramtype developed for
`flows list --orderby`, it is now relatively trivial to add in
`--orderby` support for `ls` and mark it as mutex with `--recursive`.

Tests check that the mutex setting is enforced correctly and that a
simple `--orderby` usage is encoded correctly into a query param.