Update publishing workflows
The latest guidance from the maintainers of
pypa/gh-action-pypi-publish is to split up publishing into two jobs
so as to avoid exposing the OIDC token used for Trusted Publishers to
the build process.

Switch both publishing jobs to a build-upload-download-publish flow to
achieve this separation.

Update the `docs` workflow which publishes a reference doc tarball
artifact on releases in a different way. For that, move from
`github/upload-release-asset` (deprecated) to the GitHub CLI.