Warn/hint on inactive runs when showing logs
When a `globus flows run show-logs` is run, the run will be checked to
see if its status is "INACTIVE".

This allows the CLI to print a useful hint to stderr (interactive mode
only) showing that the run log will not gain more entries until the
run is resumed.

This is an enhancement for interactive usage when a user is trying to
manually poll or "tail" the run-log and expects to see new data.
Because the log itself does not include the transition to "INACTIVE",
this usage leads to a delay between the run going inactive and the
user understanding that the state of the run has changed.

The hint is printed in a non-yellow color, and "bright_blue" was
selected from the options as a neutral color WRT the usual associated
semantics of red/yellow/green.