Add an epilog to GARE handling
CLIAuthRequirementsError can now contain a str|None named `epilog`.
When present, the handler will print it after a `* * *` separator
line. This looks similar to a textual chapter-break.

The epilog is used for a resume with session-related GARE handling to
encode a request to the user to come back with
`--skip-inactive-reason-check`.

The command series therefore will function correctly after session
"healing" performed by the user.

As part of the epilog printing, it's necessary to change the GARE
handling hook to avoid exiting eagerly if both consent and session
handling hooks return explicit exit codes. The current use-case for
explicit internal exit codes is `255`, so the logic has been selected
to ensure that a `255` is preserved.