Simplify handling of potential malformed GARE data
It turns out that the primary cause for the exit(255) condition in the
pre-existing ConsentRequired hook code is for non-GARE data. When
choosing which paths to evaluate for GARE handling, the current code
is checking `if required_scopes: ...`, meaning it is not posssible to
be running the handler with an empty list or `None`.

This effectively eliminates any concern in
`handle_internal_auth_requirements` around multiple sub-exceptions,
cascading or chaining of those errors, and reporting-up various exit
statuses requested by these components.

Instead, make the `_concrete_*_hook` methods return `None`-only, and
make it the responsibility of their callers to check for any error
conditions and exit appropriately.

For the GARE handler, this eliminates the code which had complex exit
code handling, and introduces no additional checks.
For the older consent_required_hook, the check for missing a
`required_scopes` list is moved "up one level" but its behavior is
unchanged vs `main`.