Refactor "should use map" for improved readability
The "should use map" in `WrongEntityTypeError` is used to map one
command to another, e.g. redirecting `globus endpoint update` to
`globus gcp update guest`.

This enhancement rewrites it such that the source side is expected to
be multiple (rather than the destination side), which leads to a more
natural expression of the same data.

For improved comprehensibility, the mapping is now ordered at the
top-level (as a tuple) rather than containing ordered lists, and uses
a dataclass to represent the mappings.
This makes it slightly more verbose in spite of the fact that it
features slightly less repetition of data.

A very slight data change was made to list
`globus gcs collection show` rather than `globus collection show`.