Add testing of 'collection create mapped'
This primarily consists of a new set of tests which exercise the
command exactly as written. One thing this revealed right away was
that the CLI output from mapped collection creation was missing most
of the useful fields to inspect. To remedy this, the command was
updated with the "standard_collection_fields".