Add a `globus api gcs` command
The addition of this command requires some large-diff-but-small-logic
changes to the `globus api` command definition module.

Firstly, the major logical components of these commands are separated
into three helpers:
- a decorator for the common parameters
- an "execute service command" helper
- a "handle scope string" helper (more on why this is needed below)

The command callback passes all of its parameters verbatim to the
execution helper, so there's very little change to those parts of the
command. The execution helper consumes any object of type
`globus_sdk.BaseClient`, meaning it can be passed any service client,
already authenticated/configured, and it uses that object
equivalently.

The "handle scope string" component needs to be treated differently
because it executes *before* there is a client object in hand, and
alters how that client object is authenticated for later use. It
differs subtly between the known services (handled by name) and GCS
(where `resource_server == endpoint_id`).

The GCS variant uses a positional arg for ENDPOINT_ID, so it is
declared separately from the other commands. Effectively, it's a
different interface wrapping the same functionality.