Mapped collection mutex policies use proper mutex
The MutexInfo class supports a callback for detection of the presence
of an option name or group. Use this capability to convert the
explicit in-callback checks for mutual exclusivity into a mutex
decorator which correctly describes the policy definition constraints.

Having applied this, the body logic is simple to convert into an
`if-elif-elif` chain.