Handle unrecognized errors in guest create
`globus gcs collection create guest` previously assumed that the shape
of an error from GCS met its expectations that `detail` would be dict
data. However, GCS docs do not guarantee this -- `detail` can be of
any type. Therefore, the inspection should explicitly check for this
and fail if the field is present with an unexpected type.
Failure to do so resulted in a `get()` call on a non-dict object.

Resolves #928