Improve handling of clock skew when decoding a JWT
- increase leeway from 64s to 300s
- add a warning about clock skew vs the server on failure

Regarding the increase from 64s to 300s, this represents a refinement
of the reasoning around what our tolerance should be, based on the
same relevant data + the experience of users seeing errors even in the
presence of this higher tolerance. As noted in the comments, the new
conclusion drawn is that the tolerance value selected by Kerberos, an
existing authn framework, is a reasonable default for us to use as
well. That is chosen instead of the slightly arbitrary 64s from
before.

The warning is emitted if the "Date" header in the response can be
parsed to show us a delta between the local system clock and the
server time. It is a simple echo to stderr, to guarantee that it
appears in this failure mode.

These changes have been tested manually.