Add unit tests for immature JWT signature handling
Test a number of immature signature error scenarios. One dedicated
test for the success case, and one parametrized test for a number of
"invalid parse" cases.

The positive test ensures that error parsing works normally and
produces a message. It does not inspect the error to see what the
reported clock skew is.

The failure cases revealed a bug in the `email.utils` usage.
`parsedate` returns `None` on failure, but `parsedate_to_datetime`
raises a `ValueError`. This is now handled correctly in the code which
reads the `Date` header.