Fix helptext for `--preserve-timestamp`
Transfer documentation indicates that dir mtimes are not preserved.
Because Transfer is doing async batch file processing, there are
correctness and performance issues with preserving timestamps -- it
requires an extra pass over the dir tree to retroactively set mtime on
each dir entry after the transfer completes -- and it is therefore not
included in `preserve_timestamp=True` behavior service-side.