Fix encoding of payloads as UTF-8 for `globus api`
Payloads are now encoded before being passed to the SDK, which ensures
that UTF-8 is used. This fixes an issue in which `urllib3` version 1.x
could attempt to encode data using Latin-1 and fail. The urllib3
version used is tied to the underlying supported openssl version, so
it just appears as a "platform dependent bug".

This is tricky to describe in the changelog, so it's just noted as
"encoding could fail", without full detail.