Remove autoactivate calls from the CLI
These calls to Transfer don't do anything in the post-GCSv4 universe.
Remove the helper which implemented them, and all of its call sites.

One test for handling of service-401s relied on the fact that this was
the first call to Transfer for `ls`. As a result, it had to be
rewritten -- in the course of doing so, rearrange the relevant test
data into a more self-contained test module for authentication
failures.