Add `get-subscription-info` command + share fields
This is a simpler variant of the `get-by-subscription` code which only
ever fetches the subscription info, not the underlying group. It
therefore:
- won't print a hint to stderr for text mode; if you really just
  wanted to get the group ID this is the simplest way
- has the same JSON output as `get-by-subscription`
- won't even try to fetch the group document (no network behavior)

Additionally, share some fields across commands in Groups a bit
better. `get-subscription-info` and `get-by-subscription` share
several fields, and `get-by-subscription` shares fields now with
`group show`.

As a result, there are some minor additional changes:
- `get-by-subscription` now shows the subscription ID in its text
  output
- `group show` shows subscription info *conditionally* if
  subscription_id is present and set