Combine interfaces for `display()` and `TextMode`
This is an internal change to slightly imrpove the internal API
offered by `globus_cli.termio` for the primary rendering method,
`display()`.

The enum of possible values for `display(..., text_mode=...)` are no
longer exposed directly, but instead are attributes of `display`. In
support of this, `display` is now a callable class instance, not a
simple function.

The interface change is best summarized by example:

    from globus_cli.termio import display, TextMode

    display(foo, text_mode=TextMode.text_record)

is converted to

    from globus_cli.termio import display

    display(foo, text_mode=display.RECORD)

In addition to the changes to `display`'s interface, the following
additional adjustments are made:
- `display` is "well-typed" under mypy, meaning its usages can be
  type-checked
- `TextMode`, although still used internally, is no longer exported by
  `globus_cli.termio`