Make `text_mode` a required field for `display()`
In doing so, a few adjustments are needed.
Primarily, `simple_text` has to be removed. All of its usages can be
replaced with a custom output callback, so that is done everywhere.
The reason for this removal is that `text_mode` can't meaningfully be
required when `simple_text` is available as an alternative. Requiring
it becomes awkward.

For the cases in which simple, static text is needed,
`display.static_output` is a (new) method which wraps a string as an
output callback.

All usage sites are updated as needed.