Fix a regression in handling of EPIPE
In click 7.0, a body of work introduced some specialty EPIPE handling.
At the time, we picked up the benefit and it worked under several test
scenarios. click's handling has not changed, but at some point we
updated the way we were passing exceptions to our custom excepthook.
In the course of that work, we accidentally broke the EPIPE handling,
which expects the error to ascend from `Command.invoke()` to
`Command.main()`.

We have no current need to invoke our excepthooks on OSError of any
flavor, so as the immediate fix, we can start reraising those
immediately.

In the future, we could refine this to only reraise on EPIPE,
specifically, if necessary. The unit test for this behavior specifies
EPIPE, to be robust to such an improvement.

The error can be verified and confirmed fixed with a usage like

    globus endpoint search tutorial | head -n1

Prior to this fix, broken pipe errors are sent to stderr. With the
fix, they do not appear.