Further refine error reporting for /flows/validate
This API produces non-422 errors with a details array of objects,
which is frequently overwhelming in terms of sheer information
returned.

For the textual presentation of errors, select specific fields to show
from suberror objects.

The logic runs as follows:
- if the 'detail' is an array, try to generate a better string

  - if its an array of {'loc': ..., 'msg': ...} blobs, then
    - format each element to a string
    - if there is only one, show it
    - if there are multiple, newline join them and show them

  - otherwise (not the "right shape")
    - produce the previous, compact format

This dramatically improves the presentation of validation errors
raised through our schema evaluations.